Check for file existence first
Seems silly since it's wrapped in a try / catch block, but Brian ran
into this logging an error to the console, so...🤷

+++ packages/env-loader/src/index.ts
@@ -123,9 +123,13 @@ export const processEnv = async (
 
 function cleanup() {
   try {
-    fs.unlinkSync(exitCodeFile)
+    if (fs.existsSync(buildIDFile)) {
+      fs.unlinkSync(exitCodeFile)
+    }
   } catch (e) {}
   try {
-    fs.unlinkSync(exitCodeFile)
+    if (fs.existsSync(exitCodeFile)) {
+      fs.unlinkSync(exitCodeFile)
+    }
   } catch (e) {}
 }
